fix: Alt Keyboard Shortcuts
- Common event handler for all alt shortcuts
- Move to new file
